big update
major update to Photos page which is the Arts & Craft room now. Added a grid system and moved the drawing board in the middle of it. It is now surrounded by many small grid items to be filled later. Other small changes also made including navbar name on home page